put new skills first in list, remove asterisks
If these will be first in the list on view, then that will make them very easy to find. I checked with Zemm and he agrees that this method should make them easy for the Yearn team to work with, but also these will blend in more nicely with our "normal" skills tags.